refactor(tedris): renamed tab-view and set active tab with pathname
